refactor: extract catalog domain (read + merge) into src/catalog/
